requirements.txt
asgiref==3.7.2
Django==5.0
django-jazzmin==2.6.0
django-taggit==5.0.1
Pillow==10.1.0
sqlparse==0.4.4
tzdata==2023.3
